requirements.txt
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
confection==0.1.1
contexttimer==0.3.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
decord==0.6.0
distlib==0.3.7
einops==0.6.1
executing==1.2.0
fairscale==0.4.4
fastapi==0.103.0
filelock==3.12.3
fonttools==4.42.1
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.33
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.16.4
identify==2.5.27
idna==3.4
imageio==2.31.2
importlib-metadata==6.8.0
iopath==0.1.10
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kaggle==1.5.16
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
murmurhash==1.0.9
networkx==3.1
nodeenv==1.8.0
numpy==1.25.2
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathy==0.10.2
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.16.1
portalocker==2.7.0
pre-commit==3.3.3
preshed==3.0.8
prompt-toolkit==3.0.39
protobuf==4.24.2
pure-eval==0.2.2
pyarrow==13.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pynvim==0.4.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.10.0
salesforce-lavis==1.0.2
scikit-image==0.21.0
scipy==1.11.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
streamlit==1.26.0
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
thinc==8.1.12
tifffile==2023.8.30
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.26.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.23.2
validators==0.21.2
virtualenv==20.24.4
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.48
webencodings==0.5.1
zipp==3.16.2
